em400: 32-bit add/sub

+++ em400/include/alu.h
@@ -22,7 +22,7 @@
 
 #include "registers.h"
 
-void alu_add16(int reg, uint32_t arg, uint32_t carry);
+void alu_add16(int reg, uint16_t arg, uint16_t carry);
 void alu_add32(int reg1, int reg2, uint16_t arg1, uint16_t arg2, int sign);
 void alu_mul32(int reg1, int reg2, int16_t arg);
 void alu_div32(int reg1, int reg2, int16_t arg);
